refactor "bin/upgrade list *" into one subcommand.
